// src_backup_stable/types.ts
export type CategoryType = 'IN' | 'OUT' | 'SO';
// ...
export interface ScannedItem {
  id: string;
  sku?: string;
  nama?: string;
  text?: string;
  time?: string;
  timestamp?: string;
  isCategory?: boolean;
  isLocation?: boolean;
  isInvalidSku?: boolean;
  productName?: string;
  size?: string;
  category?: CategoryType;
  kategori?: CategoryType;
  location?: string;
  lokasi?: string;
  area?: string;
  qty?: number;
  isCustomLocation?: boolean;
  [key: string]: unknown;
}